Ayy! the conex r viz
Connections are visualised and you can mouse over neurons to highlight their specific connections.

Whether or not this visualisation is actually accurate remains to be properly tested.

As in i'm pretty sure some aspect of it is brok.

Also the tragedy of having to decide between visualise and visualize for variable names only to uncover while writing this that clearly I went with the one I don't prefer.