Callback version of wire.js
Implemented a callback into the matter engine afterupdate.

This made the transfer of the navbarExpanded state work appropriately, but broke some other stuff (the plugging in)

Reverting, because we don't need the nav bar expanded state in wire js, if we can just use the position of the css transitions on the images, as those are themselves triggered by navbar expanded.

But also get the feeling that something gon fuk, so saving now.