Ok, we've messed around with repositioning salt_sim
Spent a while on repositioning, discovered that there is something fundamentally annoying going on in the physics sim on resize that stops the working model from actually working. Which is evidently v v annoying.

We can now see inside the renderer (there was some annoying faff with getting that to work this time, I think due to a refactor?). The renderer is showing messy. There is evidently some subset of wires that are not going adequatly cleaned up.

Come to think of it, that explains why there's so many < 10000 Id'd wires around.

Basically next step feels like it should be cleaning these flippers up. silly quantities of wires is definitely a candidate for the source of 'the lag', and nothing seems to be going anywhere with sorting out the resize kills network ability problem, so lets get on cleaning as a side-quest to make the main quest easier to see.